+++ java/bundles/org.eclipse.set.feature.siteplan.test/pom.xml
@@ -6,7 +6,7 @@
 	<groupId>org.eclipse.set</groupId>
 	<artifactId>org.eclipse.set.feature.siteplan.test</artifactId>
 	<version>2.6.0-SNAPSHOT</version>
-	<packaging>eclipse-plugin</packaging>
+	<packaging>eclipse-test-plugin</packaging>
 
 	<parent>
 		<relativePath>../../../</relativePath>
